new migrate float banyak stok to double

+++ database/migrations/2018_08_18_045316_create_stokbahanbakus_table.php
@@ -15,7 +15,7 @@ class CreateStokbahanbakusTable extends Migration
     {
         Schema::create('stokbahanbakus', function (Blueprint $table) {
             $table->increments('id');
-            $table->float('banyakstok');
+            $table->double('banyakstok');
             $table->string('satuan');
             $table->boolean('stokhitungluas');
             $table->integer('bahanbaku_id')->unsigned();
